refactor: languageProvider -> appLanguageProvider
