+++ packages/app-runtime/src/infrastructure/IAppLanguageProvider.ts
@@ -1,6 +1,6 @@
 import { Result } from "@js-soft/ts-utils";
 import { LanguageISO639 } from "@nmshd/core-types";
 
-export interface ILanguageProvider {
+export interface IAppLanguageProvider {
     getAppLanguage(): Promise<Result<LanguageISO639>>;
 }
