+++ packages/app-runtime/test/lib/infrastructure/FakeAppLanguageProvider.ts
@@ -1,8 +1,8 @@
 import { Result } from "@js-soft/ts-utils";
 import { LanguageISO639 } from "@nmshd/core-types";
-import { ILanguageProvider } from "../../../src";
+import { IAppLanguageProvider } from "../../../src";
 
-export class FakeLanguageProvider implements ILanguageProvider {
+export class FakeLanguageProvider implements IAppLanguageProvider {
     public getAppLanguage(): Promise<Result<LanguageISO639>> {
         return Promise.resolve(Result.ok(LanguageISO639.en));
     }
